Aggressive InPageTranslation debugging
Switch to treewalker so we can ignore elements that we do not want to translate more easily. Add missing <nav> element from list of supported elements. Ignore elementss that have a lang attribute that does not match the language we're trying to translate right now. Probably illegal to have a node filter with side effects but hey it works for now.