Experiment with HTML reunification
Try to reuse existing DOM elements in favour of inserting new ones. Hopefully, this way we don't use any javascript references or event listeners registered to the DOM nodes that are already in the document.

I've added a bit of instrumentation to see whether something ends up not making sense. So far, that is happening quite often. Not sure yet whether that's a bug in my code, whether bergamot-translator is not passing on all HTML elements, or whether the DOM changed between submitting the innerHTML for translation, and receiving the translation back.