Load models directly from translatelocally.com
Download the .tar.gz files, extract those in memory, parse the YAML file to find the filenames, and then just continue as normal. The whole finding the right file, and not actually reading the config file in detail, is pretty much a quick hack to make it just work for now.