`processedNodes` can and should be a WeakSet
If the page forgets forever about an element, so can we. Removed the WeakSet here initially because of funny business with it losing track, but looks like Firefox is behaving as expected at the moment, so daring it to be a WeakSet again. Because it could otherwise end up being a pretty massive memory leak for certain dynamic pages.