Import InPageTranslation updates (#402)
* Import InPageTranslation updates from github.com/jelmervdl/firefox-translations

* Try to make the linter happy

* Add support for translating text nodes

This seems to be necessary for supporting Youtube. It also seems to fix the issue in #385, but I still see some interaction between the Wordpress plugin and this code that should not be happening.

* Decrease translation message batch size

The changes to InPageTranslation result in at least 50% less messages per page. To make the immediately visible elements render much quicker, I reduced it further to reduce the delay.

* Reuse text nodes from page

Reuse text nodes already existing in the page if possible. This fixes issues with React pages that tend to re-add text nodes if they have been removed, showing double or untranslated text. Fixes #405.

* Improve text deletion from errornously cloned inline elements

It worked in one direction, now it works in both. Still, this scenario should almost never happen.

# Conflicts:
#	extension/view/js/InPageTranslation.js

* Let's treat `<a>` like `<span>`

# Conflicts:
#	extension/view/js/InPageTranslation.js

* Warn about text nodes that become empty after translation

# Conflicts:
#	extension/view/js/InPageTranslation.js

* Fix `removeTextNodes` call

* Make linter happy