Prevent error when text node is already parentless once updateElements() is finally called
updateElements is called on a timeout to stagger between many DOM manipulations. But in practice this could mean that we're attempting to scan a text node for translation when it has already disappeared from the document. TODO maybe we should have a general in-tree check for updateElements? I tried to cover this case with a WeakRef but weak anything doesn't seem to behave inside extensions.