Revert to using a WeakMap to track original content
This is faster and a better experience, and WeakMap was never broken to begin with. Because we only track children in entries keyed to their parents, when the reference to the parent drops, our entries with references to the children will be dropped as well. Then, that will cause those keys and their children to be dropped, etc. So that should be fine.