Replace js-untar for in-tree version
The distributed compiled version doesn't work well in an extension context as it uses blob urls to load the worker, which is not allowed in the security context. Moving it in-tree makes it easier to fix that. And I could also update the code a bit to remove some unnecessary stuff and make it a proper ES module